Drop cobra and pflag as this needs custom flag parsing to work properly
