+++ .goreleaser.yml
@@ -11,8 +11,8 @@ builds:
     binary: "py"
     ldflags:
       - -s -w
-      - -X github.com/FollowTheProcess/py/cli/cmd.version={{.Version}}
-      - -X github.com/FollowTheProcess/py/cli/cmd.commit={{.Commit}}
+      - -X github.com/FollowTheProcess/py/cli.version={{.Version}}
+      - -X github.com/FollowTheProcess/py/cli.commit={{.Commit}}
     env:
       - CGO_ENABLED=0
     goos:
@@ -33,7 +33,7 @@ brews:
       name: "Tom Fleet"
       email: "tomfleet2018@gmail.com"
     homepage: "https://github.com/FollowTheProcess/py"
-    description: "Port of the python-launcher to Go."
+    description: "Approximate port of Brett Cannon's python-launcher to Go."
     license: "MIT License"
     install: |
       bin.install "py"
