--- cli/cmd/version.go
@@ -1,25 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/fatih/color"
-	"github.com/spf13/cobra"
-)
-
-func buildVersionCmd() *cobra.Command {
-	versionCmd := &cobra.Command{
-		Use:   "version",
-		Args:  cobra.NoArgs,
-		Short: "Show py's version info",
-		Run: func(cmd *cobra.Command, args []string) {
-			ver := color.CyanString("py version")
-			sha := color.CyanString("commit")
-
-			fmt.Printf("%s: %s\n", ver, version)
-			fmt.Printf("%s: %s\n", sha, commit)
-		},
-	}
-
-	return versionCmd
-}
