+++ go.mod
@@ -1,18 +1,3 @@
 module github.com/FollowTheProcess/py
 
 go 1.17
-
-require (
-	github.com/FollowTheProcess/msg v0.2.0
-	github.com/MakeNowJust/heredoc/v2 v2.0.1
-	github.com/fatih/color v1.13.0
-	github.com/spf13/cobra v1.2.1
-)
-
-require (
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/mattn/go-colorable v0.1.11 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.0.0-20211112143042-c6105e7cf70d // indirect
-)
